telegram bot: unify message and callbackQuery handling
The old `CommandProcessor` implementation handled messages differently
to callback queries. This made it harder to keep track of longer user
interactions that used both callback queries and messages in a single
thread (e.g the `TaskCommand` when setting a task to DONE and specifying
hours taken). This refactor better deals with different update types and
allows `CommandContext` to also be supplied in callback queries. Now
both messages and callbackQueries should be handled in the
`TelegramCommand`'s `execute()` method.